Add some more utilities to the temp dir context manager

+++ structuralalignment/utils.py
@@ -5,11 +5,12 @@ import contextlib
 
 
 @contextlib.contextmanager
-def enter_temp_directory():
+def enter_temp_directory(remove=True):
     """Create and enter a temporary directory; used as context manager."""
     temp_dir = tempfile.mkdtemp()
     cwd = os.getcwd()
     os.chdir(temp_dir)
-    yield
+    yield cwd, temp_dir
     os.chdir(cwd)
-    shutil.rmtree(temp_dir)
+    if remove:
+        shutil.rmtree(temp_dir)
